Rename `lib/types` to `lib/core`
