+++ src/lib/types/codepoint-string.ts
@@ -1,6 +1,7 @@
 import { Incident } from "incident";
 import { checkedUcs2Decode } from "../_helpers/checked-ucs2-decode";
 import { lazyProperties } from "../_helpers/lazy-properties";
+import { IoType, Lazy, Reader, VersionedType, Writer } from "../core";
 import { createInvalidTypeError } from "../errors/invalid-type";
 import { createLazyOptionsError } from "../errors/lazy-options";
 import { createLowerCaseError } from "../errors/lower-case";
@@ -10,7 +11,6 @@ import { createMissingDependencyError } from "../errors/missing-dependency";
 import { createNotTrimmedError } from "../errors/not-trimmed";
 import { createPatternNotMatchedError } from "../errors/pattern-not-matched";
 import { readVisitor } from "../readers/read-visitor";
-import { IoType, Lazy, Reader, VersionedType, Writer } from "../types";
 
 let unormNfc: ((str: string) => string) | undefined = undefined;
 try {
