+++ src/lib/types/tagged-union.ts
@@ -1,10 +1,10 @@
 import { Incident } from "incident";
 import { lazyProperties } from "../_helpers/lazy-properties";
+import { IoType, Lazy, Reader, VersionedType, Writer } from "../core";
 import { createInvalidTypeError } from "../errors/invalid-type";
 import { createLazyOptionsError } from "../errors/lazy-options";
 import { createNotImplementedError } from "../errors/not-implemented";
 import { readVisitor } from "../readers/read-visitor";
-import { IoType, Lazy, Reader, VersionedType, Writer } from "../types";
 import { DocumentType } from "./document";
 import { LiteralType } from "./literal";
 import { TsEnumType } from "./ts-enum";
