+++ src/lib/types/white-list.ts
@@ -1,8 +1,8 @@
 import { Incident } from "incident";
 import { lazyProperties } from "../_helpers/lazy-properties";
+import { IoType, Lazy, Reader, VersionedType, Writer } from "../core";
 import { createLazyOptionsError } from "../errors/lazy-options";
 import { createNotImplementedError } from "../errors/not-implemented";
-import { IoType, Lazy, Reader, VersionedType, Writer } from "../types";
 
 export type Name = "white-list";
 export const name: Name = "white-list";
