+++ src/lib/writers/bson-value.ts
@@ -3,7 +3,7 @@
  */
 
 import bson from "bson";
-import { Writer } from "../types";
+import { Writer } from "../core";
 import { JsonWriter } from "./json";
 import { StructuredWriter } from "./structured";
