+++ src/lib/writers/qs-value.ts
@@ -2,7 +2,7 @@
  * @module kryo/writers/qs-value
  */
 
-import { Writer } from "../types";
+import { Writer } from "../core";
 import { JsonWriter } from "./json";
 import { StructuredWriter } from "./structured";
